Add base entity to sfr_box
